pyproject.toml
[project]
name = "project"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "fastapi[standard]>=0.115.12",
    "jupyter>=1.1.1",
    "llama-index-core>=0.12.33.post1",
    "llama-index-embeddings-huggingface>=0.5.3",
    "llama-index-llms-ollama>=0.5.4",
    "llama-index-readers-file>=0.4.7",
    "pandas>=2.2.3",
    "psycopg2-binary>=2.9.10",
    "sqlalchemy>=2.0.40",
]
